Further changes to support 16 NICs
The device distance is encoded as 3 bits per dev so we need
to change from an int to a long in order to support 16 devices

Also added example topologies to rings_test.cu

+++ src/include/rings.h
@@ -12,6 +12,6 @@ static int getDefaultThreads() {
   return ncclCudaCompCap() == 3 ? 128 : 256;
 }
 
-ncclResult_t ncclGetRings(int* nrings, int* nthreads, int rank, int nranks, int* transports, int* values, int* prev, int* next);
+ncclResult_t ncclGetRings(int* nrings, int* nthreads, int rank, int nranks, int* transports, long* values, int* prev, int* next);
 
 #endif
