+++ src/include/transport.h
@@ -47,8 +47,8 @@ struct ncclTransportComm {
 struct ncclTransport {
   const char name[4];
   ncclResult_t (*fillInfo)(ncclTinfo_t*, int);
-  ncclResult_t (*canConnect)(int*, ncclTinfo_t*, ncclTinfo_t*);
-  ncclResult_t (*getRings)(int, int*, int*, int*, int*, int*, int*, int, int*);
+  ncclResult_t (*canConnect)(long*, ncclTinfo_t*, ncclTinfo_t*);
+  ncclResult_t (*getRings)(int, int*, int*, long*, int*, int*, int*, int, int*);
   struct ncclTransportComm send;
   struct ncclTransportComm recv;
 };
